+++ KalikoCMS.Search/KalikoSearchProvider.cs
@@ -25,7 +25,7 @@ namespace KalikoCMS.Search {
 
     public class KalikoSearchProvider : SearchProviderBase {
         private readonly Collection _collection;
-        private static readonly string[] SearchFields = new[] { "title", "summary", "content", "category", "tags" };
+        private static readonly string[] SearchFields = { "title", "summary", "content", "category", "tags" };
 
         public KalikoSearchProvider() {
             _collection = new Collection("KalikoCMS");
@@ -45,13 +45,13 @@ namespace KalikoCMS.Search {
         }
 
         public override void RemoveFromIndex(Guid pageId, int languageId) {
-            string key = GetKey(pageId, languageId);
+            var key = GetKey(pageId, languageId);
             _collection.RemoveDocument(key);
         }
 
         public override void RemoveFromIndex(Collection<Guid> pageIds, int languageId) {
             foreach (var pageId in pageIds) {
-                string key = GetKey(pageId, languageId);
+                var key = GetKey(pageId, languageId);
                 _collection.RemoveDocument(key);
             }
         }
@@ -61,15 +61,15 @@ namespace KalikoCMS.Search {
         }
 
         private string GetKey(Guid pageId, int languageId) {
-            string key = string.Format("{0}:{1}", pageId, languageId);
+            var key = string.Format("{0}:{1}", pageId, languageId);
             return key;
         }
 
         public override void Init() {
-            PageFactory.PageSaved += OnPageSaved;
+            PageFactory.PagePublished += OnPagePublished;
         }
 
-        void OnPageSaved(object sender, PageEventArgs e) {
+        void OnPagePublished(object sender, PageEventArgs e) {
             IndexPage(e.Page);
         }